lang/c/msgpack: fix types
git-svn-id: file:///Users/frsyuki/project/msgpack-git/svn/x@63 5a5092ae-2292-43ba-b2d5-dcab9c1a2731

+++ msgpack/unpack/inline_context.h
@@ -42,7 +42,7 @@ typedef struct {
 typedef struct {
 	msgpack_unpack_context user;  // must be first
 	unsigned int cs;
-	size_t trail;
+	unsigned int trail;
 	unsigned int top;
 	msgpack_unpacker_stack stack[MSG_STACK_SIZE];
 } msgpack_unpacker;
